api/requirements.txt
asgiref==3.8.1
ddt==1.7.2
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
django-polymorphic==3.1.0
djangorestframework==3.15.2
factory_boy==3.3.1
Faker==35.0.0
python-dateutil==2.9.0.post0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
